requirements.txt
python-docx
pandas
fastapi
ruamel.yaml
Pillow
